Place code for page objects relating to specific pages or GUI elements inside page_objects folder
