+++ frontend/cypress/integration/actions_spec.ts
@@ -3,9 +3,9 @@ import { Progression, Priority, Role } from '../../src/api/models'
 import { FUSION_DATE_LOCALE } from '../support/helpers'
 import { barrierToString } from '../../src/utils/EnumToString'
 import { Action, Note, Participant } from '../support/mocks'
-import { ActionsGrid, ActionsTab, CreateActionDialog, EditActionDialog, mapPriority } from '../support/action'
-import { EvaluationPage, QuestionnaireSidePanel } from '../support/evaluation'
-import { ConfirmationDialog, DropdownSelect } from '../support/common'
+import { ActionsGrid, ActionsTab, CreateActionDialog, EditActionDialog, mapPriority } from '../page_objects/action'
+import { EvaluationPage, QuestionnaireSidePanel } from '../page_objects/evaluation'
+import { ConfirmationDialog, DropdownSelect } from '../page_objects/common'
 import { DELETE_ACTION } from '../support/gql'
 import * as faker from 'faker'
 import { getUsers, User } from '../support/mock/external/users'
