+++ frontend/cypress/integration/landing_page_spec.ts
@@ -3,7 +3,7 @@ import { Progression, Role } from '../../src/api/models'
 
 import { EvaluationSeed } from '../support/evaluation_seed'
 import { getUsers } from '../support/mock/external/users'
-import { EvaluationPage } from '../support/evaluation'
+import { EvaluationPage } from '../page_objects/evaluation'
 import { fusionProject1, fusionProject4 } from '../support/mock/external/projects'
 
 describe('Landing page', () => {
