+++ frontend/cypress/integration/nominations_spec.ts
@@ -1,10 +1,10 @@
 import { EvaluationSeed } from '../support/evaluation_seed'
 import { Progression, Role } from '../../src/api/models'
 import { Participant } from '../support/mocks'
-import NominationPage from '../support/nomination'
+import NominationPage from '../page_objects/nomination'
 import { getUsers } from '../support/mock/external/users'
 import * as faker from 'faker'
-import { EvaluationPage } from '../support/evaluation'
+import { EvaluationPage } from '../page_objects/evaluation'
 describe('User management', () => {
     describe('Nomination stage', () => {
         let seed: EvaluationSeed
