+++ frontend/cypress/support/common.ts
@@ -1,67 +1,3 @@
-/* It's not clear yet if common components won't clash with each other */
-export class ConfirmationDialog {
-    self = () => {
-        return cy.get('[data-testid=confirmation_dialog]')
-    }
-
-    yesButton = () => {
-        return cy.get('[data-testid=yes_button]')
-    }
-
-    noButton = () => {
-        return cy.get('[data-testid=no_button]')
-    }
-}
-
-export abstract class SideSheet {
-    /**
-     * Must indicate the top area we have control of
-     * (aka first element inside (Modal)SideSheet)
-     */
-    abstract body(): Cypress.Chainable
-
-    close = () => {
-        /* might not be stable as we can't directly grab dialog handle */
-        this.body()
-            .parent()
-            .parent()
-            .parent()
-            .within(() => {
-                cy.get('div[class*="close"]').click()
-            })
-    }
-}
-
-/**
- * Expected to work on fusion selects (elements as buttons) which appear
- * in different place in the DOM, but which all contain "section".
- * Such selects appears in different place in the dom.
- * They seem impossible to identify by any other means
- */
-export class DropdownSelect {
-    /**
-     * Selection list should already be open (one and only one)
-     *
-     * @param expected: full list of values
-     */
-    assertSelectValues = (expected: string[]) => {
-        cy.get('section').children().its('length').should('eq', expected.length)
-        expected.forEach(e => {
-            cy.get('section').contains(e).should('exist')
-        })
-    }
-
-    /**
-     * Select value with text from the list
-     * @param selectInput
-     * @param text
-     */
-    select = (selectInput: Cypress.Chainable, text: string) => {
-        selectInput.click()
-        cy.get('section').contains('span', text).click()
-    }
-}
-
 Cypress.Commands.add('getByDataTestid', (value: string, timeout: number) => {
     return cy.get(`[data-testid=${value}]`, { timeout })
 })
@@ -95,6 +31,7 @@ Cypress.Commands.add('testCacheAndDB', (testCache: Function, testDB: Function = 
     cy.reloadBmt()
     testDB()
 })
+export {}
 
 declare global {
     namespace Cypress {
@@ -107,22 +44,22 @@ declare global {
              * page reload.
              * @example cy.testCacheAndDB({assert(true)})
              */
-            testCacheAndDB(testCache: Function, testDB?: Function): Cypress.Chainable<void>
+            testCacheAndDB(testCache: Function, testDB?: Function): Cypress.Chainable
 
             /**
              * Clear + type + enter. Also works on elements clear doesn't always work on
              */
-            replace(text: string): Cypress.Chainable<Element>
+            replace(text: string): Cypress.Chainable
             /**
              * Get input for label
              */
-            for(): Cypress.Chainable<Element>
+            for(): Cypress.Chainable
             /**
              * Get element with expected data-testid. At the moment searches only globally
              * @param value Element's data-testid
              * @param timeout Max waiting time
              */
-            getByDataTestid(value: string, timeout?: number): Cypress.Chainable<Element>
+            getByDataTestid(value: string, timeout?: number): Cypress.Chainable
         }
     }
 }
